+++ R/enter_new_metadata.R
@@ -1,14 +1,13 @@
 #' Enter new metadata to fill a new scheme
 #'
-#' Open \code{emeScheme.xlsx} from system.file("googlesheet",
-#' "emeScheme.xlsx", package = "emeScheme") in excel. New data can be entered
-#' and the file has to be saved at a different location as it is a read-only
-#' file.
+#' Open \code{system.file("emeScheme.xlsx", package = "emeScheme")} in excel.
+#' New data can be entered and the file has to be saved at a different location
+#' as it is a read-only file.
 #'
 #' @param file if not \code{NULL}, the te,plate will be saved to this file.
 #' @param open if \code{TRUE}, the file will be opened. This can produce different results depending on the OS, browsr and browser settings.
 #' @param keepData if \code{TRUE} the data entry areas will be emptied. If \code{FALSE}. the example data will be included.
-#' @param format if \code{FALSE} the sheet will be opened as the googledoc sheet formated is. if \code{TRUE}, it will be formated nicely.
+#' @param format if \code{FALSE} the sheet will be opened as the sheet is. if \code{TRUE}, it will be formated nicely.
 #' @param overwrite if \code{TRUE}, the file specified in \code{file} will be overwritten. if \code{FALSE}, an error will be raised ehen the file exists.
 #' @param verbose if \code{TRUE} print usefull information
 #' @param .skipBrowseURL internal use (testing only). if \code{TRUE} skip the call of \code{browseURL()}
@@ -52,19 +51,43 @@ enter_new_metadata <- function(
     }
   )
   ##
-  fn <- file.path( tempdir(), "emeScheme.xlsx")
+
+# Warning if `format = TRUE` ----------------------------------------------
+  if (format) {
+    warning(
+    "The argument `format` is set to TRUE (the default).\n",
+    "Corruptions of the formated xlsx filer were recently observed!.\n",
+    "\n",
+    "If the resulting xlsx file is corrupt, please use\n",
+    "\n",
+    "`format = FALSE`",
+    "\n",
+    "when calling `enter_new_metadata()`\n",
+    "\n",
+    "This does NOT delete the example data.\n"
+    )
+  }
+
+# Temporary file name -----------------------------------------------------
+
+  fn <- tempfile(pattern = "emeScheme.", fileext = ".xlsx")
+
+# Format if asked for, otherwise copy to fn unchanged ---------------------
+
   if (format) {
     format_emeScheme_xlsx(
-      fn_org = system.file("googlesheet", "emeScheme.xlsx", package = "emeScheme"),
+      fn_org = system.file("emeScheme.xlsx", package = "emeScheme"),
       fn_new = fn,
       keepData = keepData
     )
   } else {
     file.copy(
-      from = system.file("googlesheet", "emeScheme.xlsx", package = "emeScheme"),
+      from = system.file("emeScheme.xlsx", package = "emeScheme"),
       to = fn
     )
   }
+
+# If file specified, copy temporary file to final destination -------------
 
   if (!is.null(file)) {
     result <- file.copy(
@@ -86,7 +109,9 @@ enter_new_metadata <- function(
       cat_ln()
     }
   }
-  ##
+
+# If open == TRUE, open the file ------------------------------------------
+
   if (open) {
     if (verbose) {
       cat_ln("Trying to open the file by opening it in the browser'", fn, "'... ")
@@ -97,7 +122,9 @@ enter_new_metadata <- function(
       utils::browseURL(fn)
     }
   }
-  ##
+
+# Return invisibly the final file name ------------------------------------
+
   invisible(fn)
 }